+++ src/test/kotlin/com/lykke/matching/engine/services/LimitOrderServiceDustTest.kt
@@ -19,7 +19,10 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Primary
 import org.springframework.test.annotation.DirtiesContext
 import org.springframework.test.context.junit4.SpringRunner
+import java.math.BigDecimal
 import kotlin.test.assertEquals
+import com.lykke.matching.engine.utils.assertEquals
+
 
 @RunWith(SpringRunner::class)
 @SpringBootTest(classes = [(TestApplicationContext::class), (LimitOrderServiceDustTest.Config::class)])
@@ -76,10 +79,10 @@ class LimitOrderServiceDustTest : AbstractTest() {
         assertEquals(OrderStatus.Matched.name, result.orders[0].order.status)
         assertEquals(OrderStatus.Processing.name, result.orders[1].order.status)
 
-        assertEquals(1000 - 0.04997355, testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
-        assertEquals(1159.92, testWalletDatabaseAccessor.getBalance("Client1", "USD"))
-        assertEquals(1000 + 0.04997355, testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
-        assertEquals(840.08, testWalletDatabaseAccessor.getBalance("Client2", "USD"))
+        assertEquals(BigDecimal.valueOf(1000 - 0.04997355), testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
+        assertEquals(BigDecimal.valueOf(1159.92), testWalletDatabaseAccessor.getBalance("Client1", "USD"))
+        assertEquals(BigDecimal.valueOf(1000 + 0.04997355), testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
+        assertEquals(BigDecimal.valueOf(840.08), testWalletDatabaseAccessor.getBalance("Client2", "USD"))
     }
 
     @Test
@@ -97,10 +100,10 @@ class LimitOrderServiceDustTest : AbstractTest() {
         assertEquals(OrderStatus.Processing.name, result.orders[0].order.status)
         assertEquals(OrderStatus.Matched.name, result.orders[1].order.status)
 
-        assertEquals(999.95, testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
-        assertEquals(1160.0, testWalletDatabaseAccessor.getBalance("Client1", "USD"))
-        assertEquals(1000.05, testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
-        assertEquals(840.0, testWalletDatabaseAccessor.getBalance("Client2", "USD"))
+        assertEquals(BigDecimal.valueOf(999.95), testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
+        assertEquals(BigDecimal.valueOf(1160.0), testWalletDatabaseAccessor.getBalance("Client1", "USD"))
+        assertEquals(BigDecimal.valueOf(1000.05), testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
+        assertEquals(BigDecimal.valueOf(840.0), testWalletDatabaseAccessor.getBalance("Client2", "USD"))
     }
 
     @Test
@@ -118,10 +121,10 @@ class LimitOrderServiceDustTest : AbstractTest() {
         assertEquals(OrderStatus.Matched.name, result.orders[0].order.status)
         assertEquals(OrderStatus.Processing.name, result.orders[1].order.status)
 
-        assertEquals(1000 + 0.04997355, testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
-        assertEquals(840.09, testWalletDatabaseAccessor.getBalance("Client1", "USD"))
-        assertEquals(1000 - 0.04997355, testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
-        assertEquals(1159.91, testWalletDatabaseAccessor.getBalance("Client2", "USD"))
+        assertEquals(BigDecimal.valueOf(1000 + 0.04997355), testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
+        assertEquals(BigDecimal.valueOf(840.09), testWalletDatabaseAccessor.getBalance("Client1", "USD"))
+        assertEquals(BigDecimal.valueOf(1000 - 0.04997355), testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
+        assertEquals(BigDecimal.valueOf(1159.91), testWalletDatabaseAccessor.getBalance("Client2", "USD"))
     }
 
     @Test
@@ -139,10 +142,10 @@ class LimitOrderServiceDustTest : AbstractTest() {
         assertEquals(OrderStatus.Processing.name, result.orders[0].order.status)
         assertEquals(OrderStatus.Matched.name, result.orders[1].order.status)
 
-        assertEquals(1000.05, testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
-        assertEquals(840.0, testWalletDatabaseAccessor.getBalance("Client1", "USD"))
-        assertEquals(999.95, testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
-        assertEquals(1160.0, testWalletDatabaseAccessor.getBalance("Client2", "USD"))
+        assertEquals(BigDecimal.valueOf(1000.05), testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
+        assertEquals(BigDecimal.valueOf(840.0), testWalletDatabaseAccessor.getBalance("Client1", "USD"))
+        assertEquals(BigDecimal.valueOf(999.95), testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
+        assertEquals(BigDecimal.valueOf(1160.0), testWalletDatabaseAccessor.getBalance("Client2", "USD"))
     }
 
     @Test
@@ -160,10 +163,10 @@ class LimitOrderServiceDustTest : AbstractTest() {
         assertEquals(OrderStatus.Matched.name, result.orders[0].order.status)
         assertEquals(OrderStatus.Processing.name, result.orders[1].order.status)
 
-        assertEquals(999.9500273, testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
-        assertEquals(1159.92, testWalletDatabaseAccessor.getBalance("Client1", "USD"))
-        assertEquals(1000.0499727, testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
-        assertEquals(840.08, testWalletDatabaseAccessor.getBalance("Client2", "USD"))
+        assertEquals(BigDecimal.valueOf(999.9500273), testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
+        assertEquals(BigDecimal.valueOf(1159.92), testWalletDatabaseAccessor.getBalance("Client1", "USD"))
+        assertEquals(BigDecimal.valueOf(1000.0499727), testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
+        assertEquals(BigDecimal.valueOf(840.08), testWalletDatabaseAccessor.getBalance("Client2", "USD"))
     }
 
     @Test
@@ -181,10 +184,10 @@ class LimitOrderServiceDustTest : AbstractTest() {
         assertEquals(OrderStatus.Matched.name, result.orders[0].order.status)
         assertEquals(OrderStatus.Processing.name, result.orders[1].order.status)
 
-        assertEquals(1000.0499727, testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
-        assertEquals(840.09, testWalletDatabaseAccessor.getBalance("Client1", "USD"))
-        assertEquals(999.9500273, testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
-        assertEquals(1159.91, testWalletDatabaseAccessor.getBalance("Client2", "USD"))
+        assertEquals(BigDecimal.valueOf(1000.0499727), testWalletDatabaseAccessor.getBalance("Client1", "BTC"))
+        assertEquals(BigDecimal.valueOf(840.09), testWalletDatabaseAccessor.getBalance("Client1", "USD"))
+        assertEquals(BigDecimal.valueOf(999.9500273), testWalletDatabaseAccessor.getBalance("Client2", "BTC"))
+        assertEquals(BigDecimal.valueOf(1159.91), testWalletDatabaseAccessor.getBalance("Client2", "USD"))
     }
 
 }
